+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.2.1] - 2021-12-05
+- fixed downloading of searcher native module release artifact
+- fixed release script
 
 ## [0.2.0] - 2021-10-11
 - replaced use of ripgrep and jq via shell command from elisp with ripgrep-backed native module