+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.3.0] - 2024-02-17
+
+## [0.3.0] - 2024-02-17
 - sort completed todos by completion date
 
 ## [0.2.2] - 2022-02-17